requirements.txt
gymnasium[atari,accept-rom-license]==0.29.1
huggingface-hub==0.17.2
hydra-core==1.3
opencv-python==4.8.0.76
pillow==10.0.0
pygame==2.5.2
torch==2.0.1
torcheval==0.0.7
tqdm==4.66.4
wandb==0.17.0
